+++ lib/lightning_web/templates/failure_notifier/failure_alert.html.heex
@@ -1,0 +1,3 @@
+<div>
+  <h1>Welcome to Sample, <%= @name %>!</h1>
+</div>
