+++ lib/lightning_web/views/failure_notifier_view.ex
@@ -1,0 +1,3 @@
+defmodule Lightning.FailureNotifierView do
+  use LightningWeb, :view
+end
