+++ mix.exs
@@ -94,7 +94,8 @@ defmodule Lightning.MixProject do
       {:telemetry_poller, "~> 1.0"},
       {:temp, "~> 0.4"},
       {:timex, "~> 3.7"},
-      {:ex_rated, "~> 2.0"}
+      {:ex_rated, "~> 2.0"},
+      {:phoenix_swoosh, "~> 1.0"}
     ]
   end
